Move Delta_ET(), Radians() and Degrees() to sun.c. Move FixAngle() and PrimeAngle() to moon.c.
